requirements.txt
ursina
Pillow